requirement.txt
absl-py==0.10.0
altair==5.3.0
blessed==1.20.0
blinker==1.8.2
cachetools==4.1.1
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.4.0
cloudpickle==1.6.0
cmake==3.26.1
contourpy==1.3.0
controlnet-aux==0.0.7
cupy-cuda12x==13.2.0
cycler==0.12.1
docker-pycreds==0.4.0
einops==0.8.0
fastrlock==0.8.2
filelock==3.10.7
fonttools==4.54.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.40
glob2==0.7
google-auth==1.22.1
google-auth-oauthlib==0.4.1
gpustat==1.1.1
grpcio==1.32.0
gym==0.17.3
gym-notices==0.0.8
h5py==3.12.1
huggingface-hub==0.23.4
idna==3.10
imageio-ffmpeg==0.5.1
importlib-metadata==6.1.0
importlib_resources==6.4.5
Jinja2==3.1.2
kiwisolver==1.4.7
lit==16.0.0
Markdown==3.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
nibabel==5.1.0
nptyping==2.5.0
numpy==1.23.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.535.108
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.1.0
opencv-python==4.6.0.66
opencv-python-headless==4.8.1.78
packaging==24.1
pathtools==0.1.2
pillow==10.4.0
pyarrow==16.1.0
pydeck==0.9.1
pydicom==2.4.4
pyglet==1.5.0
Pygments==2.18.0
pynrrd==1.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
requests==2.31.0
requests-oauthlib==1.3.0
rich==13.7.1
rsa==4.6
safetensors==0.4.3
SciencePlots==2.1.1
scipy==1.13.1
sentencepiece==0.2.0
setproctitle==1.3.3
SimpleITK==2.0.2
six==1.16.0
smmap==5.0.1
streamlit==1.36.0
streamlit-drawable-canvas==0.9.3
tenacity==8.4.2
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
timm==1.0.7
tokenizers==0.19.1
torchdiffeq==0.2.4
tqdm==4.66.5
transformers==4.42.2
typing_extensions==4.12.2
urllib3==2.2.3
wandb==0.15.4
Werkzeug==1.0.1
zipp==3.20.2
